+++ pkg/flow/mocked/mocked_flow_export_public.go
@@ -1,4 +1,4 @@
-package statuskeycardgo
+package flow
 
 import (
 	"math/rand"
