+++ pkg/flow/mocked/mocked_flow_load_account.go
@@ -1,4 +1,4 @@
-package statuskeycardgo
+package flow
 
 import (
 	"math/rand"
