+++ pkg/flow/mocked/mocked_flow_store_metadata.go
@@ -1,4 +1,4 @@
-package statuskeycardgo
+package flow
 
 import (
 	"strconv"
